+++ tests/React/Promise/FunctionAnyTest.php
@@ -2,11 +2,7 @@
 
 namespace React\Promise;
 
-/**
- * @group When
- * @group WhenAny
- */
-class WhenAnyTest extends TestCase
+class FunctionAnyTest extends TestCase
 {
     /** @test */
     public function shouldResolveToNullWithEmptyInputArray()
@@ -17,7 +13,8 @@ class WhenAnyTest extends TestCase
             ->method('__invoke')
             ->with($this->identicalTo(null));
 
-        When::any(array(), $mock);
+        any([])
+            ->then($mock);
     }
 
     /** @test */
@@ -29,10 +26,8 @@ class WhenAnyTest extends TestCase
             ->method('__invoke')
             ->with($this->identicalTo(1));
 
-        When::any(
-            array(1, 2, 3),
-            $mock
-        );
+        any([1, 2, 3])
+            ->then($mock);
     }
 
     /** @test */
@@ -44,10 +39,8 @@ class WhenAnyTest extends TestCase
             ->method('__invoke')
             ->with($this->identicalTo(1));
 
-        When::any(
-            array(When::resolve(1), When::resolve(2), When::resolve(3)),
-            $mock
-        );
+        any([resolve(1), resolve(2), resolve(3)])
+            ->then($mock);
     }
 
     /** @test */
@@ -57,13 +50,10 @@ class WhenAnyTest extends TestCase
         $mock
             ->expects($this->once())
             ->method('__invoke')
-            ->with($this->identicalTo(array(0 => 1, 1 => 2, 2 => 3)));
+            ->with($this->identicalTo([0 => 1, 1 => 2, 2 => 3]));
 
-        When::any(
-            array(When::reject(1), When::reject(2), When::reject(3)),
-            $this->expectCallableNever(),
-            $mock
-        );
+        any([reject(1), reject(2), reject(3)])
+            ->then($this->expectCallableNever(), $mock);
     }
 
     /** @test */
@@ -75,10 +65,8 @@ class WhenAnyTest extends TestCase
             ->method('__invoke')
             ->with($this->identicalTo(1));
 
-        When::any(
-            array(When::resolve(1), When::reject(2), When::reject(3)),
-            $mock
-        );
+        any([resolve(1), reject(2), reject(3)])
+            ->then($mock);
     }
 
     /** @test */
@@ -90,10 +78,8 @@ class WhenAnyTest extends TestCase
             ->method('__invoke')
             ->with($this->identicalTo(1));
 
-        When::any(
-            When::resolve(array(1, 2, 3)),
-            $mock
-        );
+        any(resolve([1, 2, 3]))
+            ->then($mock);
     }
 
     /** @test */
@@ -105,10 +91,8 @@ class WhenAnyTest extends TestCase
             ->method('__invoke')
             ->with($this->identicalTo(null));
 
-        When::any(
-            When::resolve(1),
-            $mock
-        );
+        any(resolve(1))
+            ->then($mock);
     }
 
     /** @test */
@@ -123,10 +107,8 @@ class WhenAnyTest extends TestCase
         $d1 = new Deferred();
         $d2 = new Deferred();
 
-        When::any(
-            array('abc' => $d1->promise(), 1 => $d2->promise()),
-            $mock
-        );
+        any(['abc' => $d1->promise(), 1 => $d2->promise()])
+            ->then($mock);
 
         $d2->resolve(2);
         $d1->resolve(1);