+++ tests/React/Promise/FunctionMapTest.php
@@ -2,11 +2,7 @@
 
 namespace React\Promise;
 
-/**
- * @group When
- * @group WhenMap
- */
-class WhenMapTest extends TestCase
+class FunctionMapTest extends TestCase
 {
     protected function mapper()
     {
@@ -18,7 +14,7 @@ class WhenMapTest extends TestCase
     protected function promiseMapper()
     {
         return function ($val) {
-            return When::resolve($val * 2);
+            return resolve($val * 2);
         };
     }
 
@@ -29,10 +25,10 @@ class WhenMapTest extends TestCase
         $mock
             ->expects($this->once())
             ->method('__invoke')
-            ->with($this->identicalTo(array(2, 4, 6)));
+            ->with($this->identicalTo([2, 4, 6]));
 
-        When::map(
-            array(1, 2, 3),
+        map(
+            [1, 2, 3],
             $this->mapper()
         )->then($mock);
     }
@@ -44,10 +40,10 @@ class WhenMapTest extends TestCase
         $mock
             ->expects($this->once())
             ->method('__invoke')
-            ->with($this->identicalTo(array(2, 4, 6)));
+            ->with($this->identicalTo([2, 4, 6]));
 
-        When::map(
-            array(When::resolve(1), When::resolve(2), When::resolve(3)),
+        map(
+            [resolve(1), resolve(2), resolve(3)],
             $this->mapper()
         )->then($mock);
     }
@@ -59,10 +55,10 @@ class WhenMapTest extends TestCase
         $mock
             ->expects($this->once())
             ->method('__invoke')
-            ->with($this->identicalTo(array(2, 4, 6)));
+            ->with($this->identicalTo([2, 4, 6]));
 
-        When::map(
-            array(1, When::resolve(2), 3),
+        map(
+            [1, resolve(2), 3],
             $this->mapper()
         )->then($mock);
     }
@@ -74,10 +70,10 @@ class WhenMapTest extends TestCase
         $mock
             ->expects($this->once())
             ->method('__invoke')
-            ->with($this->identicalTo(array(2, 4, 6)));
+            ->with($this->identicalTo([2, 4, 6]));
 
-        When::map(
-            array(1, 2, 3),
+        map(
+            [1, 2, 3],
             $this->promiseMapper()
         )->then($mock);
     }
@@ -89,10 +85,10 @@ class WhenMapTest extends TestCase
         $mock
             ->expects($this->once())
             ->method('__invoke')
-            ->with($this->identicalTo(array(2, 4, 6)));
+            ->with($this->identicalTo([2, 4, 6]));
 
-        When::map(
-            When::resolve(array(1, When::resolve(2), 3)),
+        map(
+            resolve([1, resolve(2), 3]),
             $this->mapper()
         )->then($mock);
     }
@@ -104,10 +100,10 @@ class WhenMapTest extends TestCase
         $mock
             ->expects($this->once())
             ->method('__invoke')
-            ->with($this->identicalTo(array()));
+            ->with($this->identicalTo([]));
 
-        When::map(
-            When::resolve(1),
+        map(
+            resolve(1),
             $this->mapper()
         )->then($mock);
     }
@@ -121,8 +117,8 @@ class WhenMapTest extends TestCase
             ->method('__invoke')
             ->with($this->identicalTo(2));
 
-        When::map(
-            array(When::resolve(1), When::reject(2), When::resolve(3)),
+        map(
+            [resolve(1), reject(2), resolve(3)],
             $this->mapper()
         )->then($this->expectCallableNever(), $mock);
     }