+++ tests/React/Promise/LazyPromiseTest.php
@@ -2,12 +2,28 @@
 
 namespace React\Promise;
 
-/**
- * @group Promise
- * @group LazyPromise
- */
 class LazyPromiseTest extends TestCase
 {
+    use PromiseTest\FullTestTrait;
+
+    public function getPromiseTestAdapter()
+    {
+        $d = new Deferred();
+
+        $factory = function () use ($d) {
+            return $d->promise();
+        };
+
+        return [
+            'promise'  => function () use ($factory) {
+                return new LazyPromise($factory);
+            },
+            'resolve'  => [$d, 'resolve'],
+            'reject'   => [$d, 'reject'],
+            'progress' => [$d, 'progress'],
+        ];
+    }
+
     /** @test */
     public function shouldNotCallFactoryIfThenIsNotInvoked()
     {
@@ -40,15 +56,15 @@ class LazyPromiseTest extends TestCase
             ->method('__invoke')
             ->will($this->returnValue(new FulfilledPromise(1)));
 
-        $fulfilledHandler = $this->createCallableMock();
-        $fulfilledHandler
+        $onFulfilled = $this->createCallableMock();
+        $onFulfilled
             ->expects($this->once())
             ->method('__invoke')
             ->with($this->identicalTo(1));
 
         $p = new LazyPromise($factory);
 
-        $p->then($fulfilledHandler);
+        $p->then($onFulfilled);
     }
 
     /** @test */
@@ -63,26 +79,26 @@ class LazyPromiseTest extends TestCase
         $p = new LazyPromise($factory);
         $this->assertInstanceOf('React\\Promise\\PromiseInterface', $p->then());
     }
-    
+
     /** @test */
     public function shouldReturnRejectedPromiseIfFactoryThrowsException()
     {
         $exception = new \Exception();
-        
+
         $factory = $this->createCallableMock();
         $factory
             ->expects($this->once())
             ->method('__invoke')
             ->will($this->throwException($exception));
 
-        $errorHandler = $this->createCallableMock();
-        $errorHandler
+        $onRejected = $this->createCallableMock();
+        $onRejected
             ->expects($this->once())
             ->method('__invoke')
             ->with($this->identicalTo($exception));
 
         $p = new LazyPromise($factory);
 
-        $p->then($this->expectCallableNever(), $errorHandler);
+        $p->then($this->expectCallableNever(), $onRejected);
     }
 }