+++ tests/React/Promise/PromiseTest/FullTestTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace React\Promise\PromiseTest;
+
+trait FullTestTrait
+{
+    use PromiseTestTrait,
+        PromiseFulfilledTestTrait,
+        PromiseRejectedTestTrait,
+        ResolveTestTrait,
+        RejectTestTrait,
+        ProgressTestTrait;
+}