+++ tests/React/Promise/RejectedPromiseTest.php
@@ -2,147 +2,43 @@
 
 namespace React\Promise;
 
-/**
- * @group Promise
- * @group RejectedPromise
- */
 class RejectedPromiseTest extends TestCase
 {
-    /** @test */
-    public function shouldReturnAPromise()
-    {
-        $p = new RejectedPromise();
-        $this->assertInstanceOf('React\\Promise\\PromiseInterface', $p->then());
-    }
+    use PromiseTest\PromiseTestTrait,
+        PromiseTest\PromiseRejectedTestTrait;
 
-    /** @test */
-    public function shouldReturnAllowNull()
+    public function getPromiseTestAdapter()
     {
-        $p = new RejectedPromise();
-        $this->assertInstanceOf('React\\Promise\\PromiseInterface', $p->then(null, null, null));
-    }
+        $val = null;
+        $promiseCalled = false;
 
-    /** @test */
-    public function shouldForwardUndefinedRejectionValue()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with(null);
+        return [
+            'promise' => function () use (&$val, &$promiseCalled) {
+                $promiseCalled = true;
 
-        $p = new RejectedPromise(1);
-        $p
-            ->then(
-                $this->expectCallableNever(),
-                function () {
-                    // Presence of rejection handler is enough to switch back
-                    // to resolve mode, even though it returns undefined.
-                    // The ONLY way to propagate a rejection is to re-throw or
-                    // return a rejected promise;
+                return new RejectedPromise($val);
+            },
+            'resolve' => function ($value) {
+                throw new \LogicException('You cannot call resolve() for React\Promise\RejectedPromise');
+            },
+            'reject' => function ($reason) use (&$val, &$promiseCalled) {
+                if ($promiseCalled) {
+                    throw new \LogicException('You must call reject() before promise() for React\Promise\RejectedPromise');
                 }
-            )
-            ->then(
-                $mock,
-                $this->expectCallableNever()
-            );
+
+                $val = $reason;
+            },
+            'progress' => function () {
+                throw new \LogicException('You cannot call progress() for React\Promise\RejectedPromise');
+            },
+        ];
     }
 
     /** @test */
-    public function shouldSwitchFromErrbacksToCallbacksWhenErrbackDoesNotExplicitlyPropagate()
+    public function shouldThrowExceptionIfConstructedWithAPromise()
     {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($this->identicalTo(2));
+        $this->setExpectedException('\InvalidArgumentException');
 
-        $p = new RejectedPromise(1);
-        $p
-            ->then(
-                $this->expectCallableNever(),
-                function ($val) {
-                    return $val + 1;
-                }
-            )
-            ->then(
-                $mock,
-                $this->expectCallableNever()
-            );
-    }
-
-    /** @test */
-    public function shouldSwitchFromErrbacksToCallbacksWhenErrbackReturnsAResolution()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($this->identicalTo(2));
-
-        $p = new RejectedPromise(1);
-        $p
-            ->then(
-                $this->expectCallableNever(),
-                function ($val) {
-                    return new FulfilledPromise($val + 1);
-                }
-            )
-            ->then(
-                $mock,
-                $this->expectCallableNever()
-            );
-    }
-
-    /** @test */
-    public function shouldPropagateRejectionsWhenErrbackThrows()
-    {
-        $exception = new \Exception();
-
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->will($this->throwException($exception));
-
-        $mock2 = $this->createCallableMock();
-        $mock2
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($this->identicalTo($exception));
-
-        $p = new RejectedPromise(1);
-        $p
-            ->then(
-                $this->expectCallableNever(),
-                $mock
-            )
-            ->then(
-                $this->expectCallableNever(),
-                $mock2
-            );
-    }
-
-    /** @test */
-    public function shouldPropagateRejectionsWhenErrbackReturnsARejection()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($this->identicalTo(2));
-
-        $p = new RejectedPromise(1);
-        $p
-            ->then(
-                $this->expectCallableNever(),
-                function ($val) {
-                    return new RejectedPromise($val + 1);
-                }
-            )
-            ->then(
-                $this->expectCallableNever(),
-                $mock
-            );
+        return new RejectedPromise(new RejectedPromise());
     }
 }